........S. [ZBX-6598] restored transaction; simplification of the new function code
git-svn-id: svn://svn.zabbix.com/branches/dev@35884 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBX-6598/src/zabbix_server/escalator/escalator.c
@@ -1341,7 +1341,12 @@ next:
 	DBfree_result(result);
 
 	/* delete completed escalations */
-	DBexecute_multiple_query("delete from escalations where", "escalationid", &escalationids);
+	if (0 != escalationids.values_num)
+	{
+		DBbegin();
+		DBexecute_multiple_query("delete from escalations where", "escalationid", &escalationids);
+		DBcommit();
+	}
 
 	zbx_free(sql);
 	zbx_vector_uint64_destroy(&escalationids);
